Moving trx burst interface to trx directory and to new 'Transceiver' cathegory in gnuradio-companion
