+++ include/grgsm/trx/CMakeLists.txt
@@ -21,20 +21,6 @@
 # Install public header files
 ########################################################################
 install(FILES
-    plotting.hpp
-    api.h
-    gsmtap.h
-    constants.h
-    gsm_constants.h
-    DESTINATION include/grgsm
+    trx_burst_if.h
+    DESTINATION include/grgsm/trx
 )
-
-add_subdirectory(decoding)
-add_subdirectory(decryption)
-add_subdirectory(demapping)
-add_subdirectory(receiver)
-add_subdirectory(misc_utils)
-add_subdirectory(qa_utils)
-add_subdirectory(flow_control)
-add_subdirectory(transmitter)
-add_subdirectory(trx)
