+++ lib/trx/CMakeLists.txt
@@ -17,24 +17,7 @@
 # the Free Software Foundation, Inc., 51 Franklin Street,
 # Boston, MA 02110-1301, USA.
 
-########################################################################
-# Install public header files
-########################################################################
-install(FILES
-    plotting.hpp
-    api.h
-    gsmtap.h
-    constants.h
-    gsm_constants.h
-    DESTINATION include/grgsm
+add_sources(
+    trx_burst_if_impl.cc
 )
 
-add_subdirectory(decoding)
-add_subdirectory(decryption)
-add_subdirectory(demapping)
-add_subdirectory(receiver)
-add_subdirectory(misc_utils)
-add_subdirectory(qa_utils)
-add_subdirectory(flow_control)
-add_subdirectory(transmitter)
-add_subdirectory(trx)
